Name change, tokenizer extracted
Gem constant changed from Differ to Diffident and files updated.

Tokenizer extracted from main file. Decoupling in preparation for
StringScan rewrite.

Spec helper updated to default to colored output.

Responsibility for keeping track of if the separator is printable or
not moved from main initializer to Diff that actually uses it.